fix(lint): Remove unnecessary uint32 conversions and preallocate slice
- Remove unnecessary uint32() conversions in calls.go, closures.go, statements.go
  (StartPoint().Row already returns uint32)
- Preallocate vars slice with capacity in variables.go for better performance
- Fixes all remaining golangci-lint issues (unconvert, prealloc)

Verification:
- golangci-lint run passes (0 issues)
- go test ./graph/golang/... passes

+++ sast-engine/graph/golang/calls.go
@@ -23,7 +23,7 @@ func ParseCallExpression(node *sitter.Node, sourceCode []byte) *CallInfo {
 	}
 
 	info := &CallInfo{
-		LineNumber: uint32(node.StartPoint().Row) + 1,
+		LineNumber: node.StartPoint().Row + 1,
 		StartByte:  node.StartByte(),
 		EndByte:    node.EndByte(),
 	}
